+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/listeners/VcsMappingChangedListener.kt
@@ -29,7 +29,7 @@ class VcsMappingChangedListener(val project: Project) : VcsRepositoryMappingList
 
         projectSettings.gitlabUrls = currentGitlabUrls
         AppSettings.instance.remotes
-            .mapNotNull { it.value.gitlabUrl }
+            .mapNotNull { it.value?.gitlabUrl }
             .forEach { projectSettings.gitlabUrls += it }
 
         if (projectSettings.newProject) {