+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/settings/AppSettings.kt
@@ -37,7 +37,7 @@ class AppSettings : PersistentStateComponent<AppSettings> {
     var hits = 0
     var requestSupport = true
     // Remote url to Remote
-    var remotes: MutableMap<String, Remote> = mutableMapOf()
+    var remotes: MutableMap<String, Remote?> = mutableMapOf()
     var lintFrequency: LintFrequencyEnum = LintFrequencyEnum.ON_SAVE
     var showMergedPreview = true
     var allowSelfSignedCertificate = false